[VL] Extract ORC reader compilation from GenericBenchmark (#3414)
