+++ cpp/velox/benchmarks/common/BenchmarkUtils.cc
